% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_then_install}
\alias{check_then_install}
\title{Modified install package}
\usage{
check_then_install(pkg, ...)
}
\arguments{
\item{pkg}{A packages to install}

\item{...}{arguments passed to install.packages()}
}
\description{
Checks whether the list of packages is installed. If not, then it installs it.
}
