% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dens_at_grid}
\alias{dens_at_grid}
\title{Evaluate density at user-specified grid}
\usage{
dens_at_grid(point, data, ...)
}
\arguments{
\item{point}{grid of points at which to evaluate the density of data.}

\item{data}{data with which to estimate the density.}

\item{...}{arguments passed to density()}
}
\description{
Estimates the density at a grid of points specified by the user
}
