% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{g_legend}
\alias{g_legend}
\title{Extract legend from ggplot}
\usage{
g_legend(theplot)
}
\arguments{
\item{theplot}{A ggplot2 object.}
}
\description{
This function takes a ggplot2 object and extracts the legend as a grob.
}
