% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{position_dodgev}
\alias{position_dodgev}
\title{Adjust position by dodging overlaps to the side. All code written by Jared Lander available from 
https://github.com/jaredlander/coefplot/blob/master/R/position.r/}
\usage{
position_dodgev(height = NULL)
}
\arguments{
\item{height}{Dodging height, when different to the height of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples for a use case.}
}
\description{
Adjust position by dodging overlaps to the side. All code written by Jared Lander available from 
https://github.com/jaredlander/coefplot/blob/master/R/position.r/
}
