% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vcovCluster}
\alias{vcovCluster}
\title{Cluster-robust standard errors}
\usage{
vcovCluster(model, cluster)
}
\arguments{
\item{model}{An estimated regression model from lm()}

\item{cluster}{A variable that indicates which cluster each observation belongs to.}
}
\description{
Computes cluster-robust standard errors, with automatic missing data handling.
}
